Add files via upload
Added a function to remove any mft_classification that resulted in a 'not' value.  Josh's code gave a probability if a text was and was not a specific mft.  Removing them the probability of 'not_mft' cleaned up errors.  I'm not testing if mft isn't present but if it is.